Externalise JeedomMock classes
