+++ tests/RunVeoliaTestsSimple.sh
@@ -6,4 +6,4 @@ cp ./veolia_sudest_data_src/veolia_html_4March.htm veolia_sudest_data
 cp ./veolia_sudest_data_src/veolia_html_nodata.htm veolia_sudest_data
 cp ./veolia_sudest_data_src/veolia_html_Feb18.htm veolia_sudest_data
 cp ./veolia_sudest_data_src/veolia_html_Feb18_Non_Mesure.htm veolia_sudest_data
-php testVeoliaSudEstSimple.php |grep -v "It is not safe to rely on the system"
+php testVeoliaSudEstSimple.php
