+++ tests/testVeoliaSudEst.php
@@ -3,120 +3,8 @@
  *
  */
 
+require_once './veolia_eauMockJeedom.class.php';
 require_once '../core/class/veolia_eau_process.class.php' ;
-
-class eqLogic{
-  private $website = 2;
-  private $login = 1;
-  private $password = "PSW1";
-  private $depart = 0;
-  private $compteur = 0;
-  private $last="2018-03-01";
-  private $alert="#680";
-  private $mock_file="veolia_sudest_data_src/veolia_html_3March.htm";
-  private $mock_date="2018-03-07";
-  private $maxday=1000;
-  private $maxmonth=10000;
-  public function getConfiguration($config){
-    if ($config=="website"){
-      return $this->website;
-    } elseif ($config=="login"){
-      return $this->login;
-    } elseif ($config=="password"){
-      return $this->password;
-    }elseif ($config=="depart"){
-      return $this->depart;
-    }elseif ($config=="compteur"){
-      return $this->compteur;
-    }elseif ($config=="last"){
-      return $this->last;
-    }elseif ($config=="alert"){
-      return $this->alert;
-    }elseif ($config=="mock_file"){
-      return $this->mock_file;
-    }elseif ($config=="mock_date"){
-      return $this->mock_date;
-    }elseif ($config=="maxday"){
-      return $this->maxday;
-    }elseif ($config=="maxmonth"){
-      return $this->maxmonth;
-    }
-    else {
-      log::add('veolia_eau','debug','getConfiguration:'.$config);
-      return "ko";
-    }
-  }
-  public function setConfiguration($config, $value){
-      if ($config=="website"){
-        $this->website=$value;
-      } elseif ($config=="login"){
-        $this->login=$value;
-      } elseif ($config=="password"){
-        $this->password=$value;
-      }elseif ($config=="depart"){
-        $this->depart=$value;
-      }elseif ($config=="compteur"){
-        $this->compteur=$value;
-      }elseif ($config=="last"){
-        $this->last=$value;
-      }elseif ($config=="alert"){
-        $this->alert=$value;
-      }elseif ($config=="mock_file"){
-        $this->mock_file=$value;
-      }elseif ($config=="mock_date"){
-        $this->mock_date=$value;
-      }
-      else {
-        log::add('veolia_eau','debug','setConfiguration:'.$config." ".$value);
-        return "ko";
-      }
-  }
-  public function displayConfig(){
-         log::add('veolia_eau','debug','displayConfig:'.
-         " website:".
-          $this->website.
-          " login:".
-          $this->login.
-          " password:".
-          $this->password.
-          " depart:".
-          $this->depart.
-          " compteur:".
-          $this->compteur.
-          " last:".
-          $this->last.
-          " alert:".
-          $this->alert.
-          " mock_date:".
-          $this->mock_date
-          );
-  }
-  function getCmd($a,$b)
-  {
-    return 0;
-  }
-  function save ($cmd){
-    // log::add('veolia_eau','debug','save:'.$cmd);
-  }
-}
-class cmd {
-  static public function byId($argument)
-  {
-    return new cmd;
-  }
-  public function execCmd($cmd)
-  {
-    log::add('veolia_eau','debug','execCmd:'.serialize($cmd));
-  }
-}
-
-class log {
-  static public function add($plugin,$level,$message)
-  {
-    echo "\n".$plugin." ".$level." ".$message."\n";
-  }
-}
-
 
 // Unit tests //
 //class veoliaTest extends PHPUnit_Framework_TestCase {
@@ -131,6 +19,7 @@ class log {
 //  }
 //}
 $instanceVeolia = new veolia_eau;
+$instanceVeolia->setConfiguration('last',"2018-03-01");
 $instanceVeolia->setConfiguration('mock_date',"2018-03-07");
 $instanceVeolia->setConfiguration('mock_file',"veolia_sudest_data/veolia_html_3March.htm");
 $instanceVeolia->displayConfig();
@@ -186,8 +75,6 @@ $instanceVeolia->displayConfig();
 //$instanceVeolia->setConfiguration('password',"xx");
 //$instanceVeolia->getConso(1);
 //$instanceVeolia->displayConfig();
-
-
 
 /**
  *
